Module mathml::ast

Modules

operator

Structs

Ci

Represents content identifiers such that variables can be have type attribute to specify what it represents. E.g. function, vector, real,…

Differential

Represents the differentiation operator diff for functions or expresions func E.g. df/dx

DownArrow

Handles ↓ such that {comp}↓_{sub}^{sup} can parse

HatComp

Represents Hat operator and the contents the hat operator is being applied to. E.g. f \hat{x}, where op will be \hat{x}, comp with be the contents

Integral

Integral operation represents integral over math expressions

LaplacianComp

Laplacian operator of vector calculus which takes in argument comp. E.g. ∇^2 f

Math

The Math struct represents the corresponding element type in MathML 3 (https://www.w3.org/TR/MathML3/appendixa.html#parsing_math)

Mi

Represents identifiers such as variables, function names, constants. E.g. x , sin, n

Mrow

Represents groups of any subexpressions E.g. For a given expression: 2x+4 –> 2 and x are grouped together

SummationMath

Represents sum operator with terms that are summed over E.g. sum_{i=1}^{K} f(x_i)

Enums

MathExpression

The MathExpression enum is not faithful to the corresponding element type in MathML 3 (https://www.w3.org/TR/MathML3/appendixa.html#parsing_MathExpression)

Type