Module mathml::expression

Structs

Expression

PreExp

Intermediate data structure to support the generation of graphs of mathematical expressions

Enums

Atom

Expr

Functions

all_ops_are_add_or_subt

Check if the current term’s operators are all add or subtract.

all_ops_are_mult_or_div

Check if the current term’s operators are all multiply or divide.

contains_redundant_parens

The graph generation process adds additional parentheses to preserve operation order. This function checks for the existence of those redundant parentheses.

get_node_idx

Return the node index if it exists; if not, add a new node and return the node index.

is_derivative

Check if the fraction is a derivative expressed in Leibniz notation. If yes, mutate it to remove the ‘d’ prefixes.

need_to_distribute

Check if the current term’s operators contain add or minus and without the unary operator.

need_to_distribute_divs

Check if the current term’s operators contain multiple divisions, and the denominators contain add and subtract and without unary operators

ops_contain_mult

Check if the current term’s operators contain multiply.

preprocess_content

Preprocess the content prior to parsing.

remove_redundant_mrow

Remove redundant mrow next to specific MathML elements. This function will likely be removed once the img2mml pipeline is fixed.

remove_redundant_mrows

Remove redundant mrows in mathml because some mathml elements don’t need mrow to wrap. This function will likely be removed once the img2mml pipeline is fixed.

remove_redundant_parens

Remove redundant parentheses.

switch_add_subt

Switch the summation operator and the subtraction operator

switch_mul_div

Switch the multiplication operator and the division operator

wrap_math

Wrap mathml vectors by mrow as a single expression to process

Type Definitions

MathExpressionGraph

Struct for representing mathematical expressions in order to align with source code.