Module mathml::parsers::generic_mathml

Structs

ParseError

Traits

FromFile

Functions

add

attribute

equals

lparen

math_expression

Math expressions

mfrac

Fractions

mi

Identifiers

mn

Numbers

mo

Operators

mover

mrow

Rows

msqrt

Square roots

msub

Subscripts

msubsup

msup

Superscripts

operator

parse

The parse function is part of the public API. It takes a string and returns a Math object.

quoted_string

Quoted string

rparen

subtract

ws

A combinator that takes a parser inner and produces a parser that also consumes both leading and trailing whitespace, returning the output of inner.

xml_declaration

Type Definitions

IResult

Redefine IResult, filling in the first generic type parameter with Span, for increased brevity.

Span