Module mathml::parsers::interpreted_mathml

This module contains parsers that perform some amount of preliminary domain-specific interpretation of presentation MathML (e.g., globbing S(t) to an identifier S of type ‘function’). This is in contrast to the generic_mathml.rs module that contains parsers that do not attempt to perform any interpretation but instead simply preserve the original MathML document structure.

Functions

ci_subscript

Parse content identifier for Msub

ci_univariate_func

Parse content identifiers corresponding to univariate functions. Example: S(t)

ci_unknown

Parse a content identifier of unknown type.

first_order_derivative_leibniz_notation

Parse a first-order ordinary derivative written in Leibniz notation.

interpreted_math

Parser for interpreted math expressions. testing MathML documents

math_expression

Parser for math expressions. This varies from the one in the generic_mathml module, since it assumes that expressions such as S(t) are actually univariate functions.

mfrac

mrow

newtonian_derivative

operator

Function to parse operators. This function differs from the one in parsers::generic_mathml by disallowing operators besides +, -, =, (, and ).